insta/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
black==22.10.0
click==8.1.3
colorama==0.4.5
Django==3.2.12
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-imagekit==4.1.0
mypy-extensions==0.4.3
pathspec==0.10.1
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2021.5
